Renamed project to Micro RTPS-DDS Gen (#12)
* Removed unused previous tests

* Refs #3451. Updated with MicroXRCE and MicroCDR name changes.

* Refs #3451. Updated Product name. Works with new repos github address.

* Added argument to select client's branch.
