+++ README.md
@@ -1,12 +1,12 @@
-# eProsima Micro RTPS
+# eProsima Micro XRCE
 
 <a href="http://www.eprosima.com"><img src="https://encrypted-tbn3.gstatic.com/images?q=tbn:ANd9GcSd0PDlVz1U_7MgdTe0FRIWD0Jc9_YH-gGi0ZpLkr-qgCI6ZEoJZ5GBqQ" align="left" hspace="8" vspace="2" width="100" height="100" ></a>
 
-*eProsima Micro RTPS* is a software solution which allows to communicate eXtremely Resource Constrained Environments (XRCEs) with an existing DDS network. This implementation complies with the specification proposal, "eXtremely Resource Constrained Environments DDS (DDS-XRCE)" submitted to the Object Management Group (OMG) consortium.
+*eProsima Micro XRCE* is a software solution which allows to communicate eXtremely Resource Constrained Environments (XRCEs) with an existing DDS network. This implementation complies with the specification proposal, "eXtremely Resource Constrained Environments DDS (DDS-XRCE)" submitted to the Object Management Group (OMG) consortium.
 
-*Micro RTPS* implements a client-server protocol to enable resource-constrained devices (clients) to take part in DDS communications. *Micro RTPS Agent* (server) makes possible this communication. The *Micro RTPS Agent* acts on behalf of the *Micro RTPS Clients* and enables them to take part as DDS publishers and/or subscribers in the DDS Global Data Space.
+*Micro XRCE* implements a client-server protocol to enable resource-constrained devices (clients) to take part in DDS communications. *Micro XRCE Agent* (server) makes possible this communication. The *Micro XRCE Agent* acts on behalf of the *Micro XRCE Clients* and enables them to take part as DDS publishers and/or subscribers in the DDS Global Data Space.
 
-*Micro RTPS* provides both, a plug and play *Micro RTPS Agent* and an API layer which allows you to implement your *Micro RTPS Clients*.
+*Micro XRCE* provides both, a plug and play *Micro XRCE Agent* and an API layer which allows you to implement your *Micro XRCE Clients*.
 
 ![Architecture](docs/architecture.png)
 
@@ -14,9 +14,9 @@
 
 You can access the documentation online, which is hosted on Read the Docs.
 
-* [Start Page](http://micro-rtps.readthedocs.io)
-* [Installation manual](http://micro-rtps.readthedocs.io/en/latest/installation.html)
-* [User manual](http://micro-rtps.readthedocs.io/en/latest/introduction.html)
+* [Start Page](http://micro-xrce.readthedocs.io)
+* [Installation manual](http://micro-xrce.readthedocs.io/en/latest/installation.html)
+* [User manual](http://micro-xrce.readthedocs.io/en/latest/introduction.html)
 
 ## Getting Help
