--- scripts/micrortpsgen.bat
@@ -1,30 +1,0 @@
-@echo off
-setlocal
-
-set dir=%~dp0
-set args=%1
-
-
-:getarg
-shift
-if "%~1"=="" goto continue
-set args=%args% %1
-goto getarg
-
-:continue
-
-:: Check java binary.
-set java_exec=java
-
-java -version > NUL 2>&1
-
-if not %ERRORLEVEL%==0 (
-   if not "%JAVA_HOME%"=="" (
-      set java_exec="%JAVA_HOME%\bin\java"
-   ) else (
-      echo Java binary cannot be found. Please, make sure its location is in the PATH environment variable or set JAVA_HOME environment variable.
-      exit /B 65
-   )
-)
-
-%java_exec% -jar "%dir%\..\share\micrortps\micrortpsgen.jar" %args%
