+++ src/main/resources/com/eprosima/uxr/idl/templates/Common.stg
