+++ src/main/resources/com/eprosima/uxr/idl/templates/PublisherSource.stg
@@ -1,0 +1,140 @@
+// Copyright 2016 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+group ProtocolHeader;
+
+main(ctx, definitions) ::= <<
+$fileHeader(ctx=ctx,  file=[ctx.filename, "Publisher.c"], description=["This file contains the implementation of the publisher functions."])$
+
+#include "$ctx.filename$.h"
+
+#include <uxr/client/client.h>
+#include <ucdr/microcdr.h>
+
+#include <stdio.h>
+
+#define STREAM_HISTORY  8
+#define BUFFER_SIZE     UXR_CONFIG_UDP_TRANSPORT_MTU * STREAM_HISTORY
+
+$definitions; separator="\n"$
+
+static const char* topic_xml = "<dds><topic>$if(ctx.lastStructure.typeCode.hasKey)$<kind>WITH_KEY</kind>$endif$<name>$ctx.filename$PubSubTopic</name><dataType>$ctx.lastStructure.scopedname$</dataType></topic></dds>";
+static const char* publisher_xml = "<publisher name=\"MyPublisher\"";
+static const char* datawriter_xml = "<profiles><publisher profile_name=\"default_xrce_publisher_profile\"><topic><kind>$if(ctx.lastStructure.typeCode.hasKey)$WITH_KEY$else$NO_KEY$endif$</kind><name>$ctx.filename$PubSubTopic</name><dataType>$ctx.lastStructure.scopedname$</dataType><historyQos><kind>KEEP_LAST</kind><depth>5</depth></historyQos><durability><kind>TRANSIENT_LOCAL</kind></durability></topic></publisher></profiles>";
+
+int main(void)
+{
+    // Transport
+    uxrUDPTransport transport;
+    if(!uxr_init_udp_transport(&transport, "127.0.0.1", 2018))
+    {
+        printf("Error at create transport.\n");
+        return 1;
+    }
+
+    // Session
+    uxrSession session;
+    uxr_init_session(&session, &transport.comm, 0xAAAABBBB);
+    if(!uxr_create_session(&session))
+    {
+        printf("Error at create session.\n");
+        return 1;
+    }
+
+    // Streams
+    uint8_t output_reliable_stream_buffer[BUFFER_SIZE];
+    uxrStreamId reliable_out = uxr_create_output_reliable_stream(&session, output_reliable_stream_buffer, BUFFER_SIZE, STREAM_HISTORY);
+
+    uint8_t input_reliable_stream_buffer[BUFFER_SIZE];
+    uxr_create_input_reliable_stream(&session, input_reliable_stream_buffer, BUFFER_SIZE, STREAM_HISTORY);
+
+    // Create entities
+    uxrObjectId participant_id = uxr_object_id(0x01, UXR_PARTICIPANT_ID);
+    uint16_t participant_req = uxr_write_create_participant_ref(&session, reliable_out, participant_id, 0, "default participant", UXR_REPLACE);
+
+    uxrObjectId topic_id = uxr_object_id(0x01, UXR_TOPIC_ID);
+    uint16_t topic_req = uxr_write_configure_topic_xml(&session, reliable_out, topic_id, participant_id, topic_xml, UXR_REPLACE);
+
+    uxrObjectId publisher_id = uxr_object_id(0x01, UXR_PUBLISHER_ID);
+    uint16_t publisher_req = uxr_write_configure_publisher_xml(&session, reliable_out, publisher_id, participant_id, publisher_xml, UXR_REPLACE);
+
+    uxrObjectId datawriter_id = uxr_object_id(0x01, UXR_DATAWRITER_ID);
+    uint16_t datawriter_req = uxr_write_configure_datawriter_xml(&session, reliable_out, datawriter_id, publisher_id, datawriter_xml, UXR_REPLACE);
+
+    // Send create entities message and wait its status
+    uint8_t status[4];
+    uint16_t requests[4] = {participant_req, topic_req, publisher_req, datawriter_req};
+    if(!uxr_run_session_until_all_status(&session, 1000, requests, status, 4))
+    {
+        printf("Error at create entities: participant: %i topic: %i publisher: %i darawriter: %i\n", status[0], status[1], status[2], status[3]);
+        return 1;
+    }
+
+    // Write topics
+    while(true)
+    {
+        $ctx.lastStructure.cScopedname$ topic = {0};
+
+        ucdrBuffer mb;
+        uint32_t topic_size = $ctx.lastStructure.cScopedname$_size_of_topic(&topic, 0);
+        uxr_prepare_output_stream(&session, reliable_out, datawriter_id, &mb, topic_size);
+        $ctx.lastStructure.cScopedname$_serialize_topic(&mb, &topic);
+
+        uxr_run_session_until_timeout(&session, 1000);
+    }
+
+    // Delete resources
+    uxr_delete_session(&session);
+
+    return 0;
+}
+>>
+
+module(ctx, parent, module, definition_list) ::= <<
+$definition_list$
+>>
+
+definition_list(definitions) ::= <<
+$definitions; separator="\n"$
+
+>>
+
+annotation(ctx, annotation) ::= <<>>
+
+interface(ctx, parent, interface, export_list) ::= <<
+$export_list$
+>>
+
+export_list(exports) ::= <<
+$exports; separator="\n"$
+
+>>
+
+exception(ctx, parent, exception) ::= <<>>
+
+operation(ctx, parent, operation, param_list) ::= <<>>
+
+param_list(parameters) ::= <<>>
+
+param(parameter) ::= <<>>
+
+const_decl(ctx, parent, const) ::= <<>>
+
+typedef_decl(ctx, parent, typedefs) ::= <<>>
+
+struct_type(ctx, parent, struct) ::= <<>>
+
+union_type(ctx, parent, union) ::= <<>>
+
+enum_type(ctx, parent, enum) ::= <<>>
