--- utils/releaseTest/.gitignore
@@ -1,4 +1,0 @@
-*.h
-*.c
-CMakeLists.txt
-build
