--- utils/releaseTest/IncludeIDL/IncludeIDL1.idl
@@ -1,4 +1,0 @@
-struct S1
-{
-	long value;
-};
