--- utils/releaseTest/MultipleIDL/MultipleIDL1.idl
@@ -1,6 +1,0 @@
-struct MYSTRUCT1
-{
-	long value;
-	string message;
-};
-
