--- utils/releaseTest/MultipleIDL/MultipleIDL2.idl
@@ -1,5 +1,0 @@
-struct MYSTRUCT2
-{
-	long value;
-	string message;
-};
