--- utils/releaseTest/SimpleIDL/SimpleIDL.idl
@@ -1,18 +1,0 @@
-struct Basics
-{
-	boolean var_boolean;
-	char var_char;
-    octet var_octet;
-
-	short ashort;
-    unsigned short var_ushort;
-
-	long var_long;
-	unsigned long var_ulong;
-
-	long long var_longlong;
-	unsigned long long var_ulonglong;
-
-	float var_float;
-	double var_double;
-};
